Updated split_barcodes to get a substring first and rev-comp afterwards
This was the original discussion:

>> Hi Felix,
 
Regards the index issue with 5557 (I needed to trim the last base of the index and add a C to the start for it to work). we’ve realised this might be due to the way the i5 index is processed on your end? i.e. it trims the first base then does revcomp (or it does revcomp then trims the last base) when it should be trimming the opposite.
 
Does this make sense?
 
I don’t need anything re-processing as the fix worked fine, but I want to avoid this issue in the future.
 
Cheers,
 
Stephen 
 
From: "ngs.info" <ngs.info@babraham.ac.uk>
Date: Thursday, 9 September 2021 at 10:18
To: Stephen Clark <Stephen.Clark@babraham.ac.uk>, "ngs.info" <ngs.info@babraham.ac.uk>
Subject: RE: QC for 5557
 
Yes it’s your bioinformatics but I think I might have caused the issue by doing a 9 cycle read rather than the 8 we usually do for Nextera. I’m wondering if the pipeline is set to keep the last 8 bases of the read or something.
 
Sorry for causing you issues, 
 
Milly